# services/core/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud DevBox API
  description: |
    Cloud DevBox 是一个云端开发环境平台，提供完整的 RESTful API 用于管理开发环境、模板、部署、协作等功能。
    
    ## 认证方式
    
    API 支持两种认证方式：
    
    1. **JWT Token** - 用于 Web 应用和桌面客户端
       - 通过 `/api/v1/auth/login` 获取 token
       - 在请求头中添加 `Authorization: Bearer <token>`
    
    2. **API Key** - 用于程序化访问和 CI/CD 集成
       - 通过 Web 控制台或 API 创建 API Key
       - 在请求头中添加 `X-API-Key: <api_key>`
    
    ## 速率限制
    
    - 默认限制：1000 次/小时/用户
    - 超出限制返回 429 状态码
    - 响应头包含 `X-RateLimit-Remaining` 和 `X-RateLimit-Reset`
    
    ## 错误处理
    
    所有错误响应遵循统一格式：
    ```json
    {
      "code": "ERROR_CODE",
      "message": "Human readable message",
      "details": {},
      "request_id": "uuid",
      "timestamp": "2026-01-13T00:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: Cloud DevBox Team
    email: support@clouddevbox.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.clouddevbox.io/api/v1
    description: Production server
  - url: https://staging-api.clouddevbox.io/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api-keys:
    get:
      tags: [API Keys]
      summary: 列出 API 密钥
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAPIKeys'
    post:
      tags: [API Keys]
      summary: 创建 API 密钥
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
...
  /api-keys/{id}:
    get:
      tags: [API Keys]
      summary: 获取 API 密钥详情
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
    put:
      tags: [API Keys]
      summary: 更新 API 密钥
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAPIKeyRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
    delete:
      tags: [API Keys]
      summary: 删除 API 密钥
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      responses:
        '204':
          description: 删除成功
  /api-keys/{id}/revoke:
    post:
      tags: [API Keys]
      summary: 撤销 API 密钥
      parameters:
        - $ref: '#/components/parameters/APIKeyId'
      responses:
        '200':
          description: 撤销成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ============================================================================
  # Webhooks
  # ============================================================================
  /webhooks:
    get:
      tags: [Webhooks]
      summary: 列出 Webhook
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWebhooks'
    post:
      tags: [Webhooks]
      summary: 创建 Webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
...
  /webhooks/{id}:
    get:
      tags: [Webhooks]
      summary: 获取 Webhook 详情
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
    put:
      tags: [Webhooks]
      summary: 更新 Webhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
    delete:
      tags: [Webhooks]
      summary: 删除 Webhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '204':
          description: 删除成功
  /webhooks/{id}/active:
    put:
      tags: [Webhooks]
      summary: 设置 Webhook 激活状态
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                active:
                  type: boolean
              required: [active]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /webhooks/{id}/secret:
    post:
      tags: [Webhooks]
      summary: 重新生成 Webhook 密钥
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
  /webhooks/{id}/test:
    post:
      tags: [Webhooks]
      summary: 测试 Webhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: 测试发送成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    APIKeyId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WebhookId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
          format: int64
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean

    # ============================================================================
    # Authentication Schemas
    # ============================================================================
...
    APIKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        key_prefix:
          type: string
          description: API Key 前缀（用于识别）
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
        last_used_ip:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        scopes:
          type: array
          items:
            type: string
          minItems: 1
        expires_at:
          type: string
          format: date-time
      required: [name, scopes]
    CreateAPIKeyResponse:
      type: object
      properties:
        api_key:
          $ref: '#/components/schemas/APIKey'
        plain_key:
          type: string
          description: 完整的 API Key（仅在创建时返回一次）
    UpdateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        scopes:
          type: array
          items:
            type: string
    PaginatedAPIKeys:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/APIKey'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ============================================================================
    # Webhook Schemas
    # ============================================================================
    Webhook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        headers:
          type: object
          additionalProperties:
            type: string
        retry_config:
          $ref: '#/components/schemas/WebhookRetryConfig'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
          minItems: 1
        headers:
          type: object
          additionalProperties:
            type: string
        retry_config:
          $ref: '#/components/schemas/WebhookRetryConfig'
      required: [name, url, events]
    WebhookRetryConfig:
      type: object
      properties:
        max_retries:
          type: integer
          default: 3
        retry_delay_ms:
          type: integer
          format: int64
          default: 1000
        backoff_factor:
          type: number
          format: double
          default: 2.0
        max_delay_ms:
          type: integer
          format: int64
          default: 60000
...
    PaginatedWebhooks:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Webhook'
        pagination:
          $ref: '#/components/schemas/Pagination'
...